pyproject.toml
[project]
name = "ai-file-search"
version = "0.1.0"
description = "Local AI file-search"
package-mode = false
authors = [
    {name = "Anson Lam",email = "ziyilam@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "pdfminer-six (>=20250506,<20250507)",
    "python-docx (>=1.2.0,<2.0.0)",
    "sentence-transformers (>=5.0.0,<6.0.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "watchdog (>=6.0.0,<7.0.0)",
    "llama-cpp-python (>=0.3.12,<0.4.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "psutil (>=7.0.0,<8.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "typer[all] (>=0.16.0,<0.17.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "flask (>=3.0.0,<4.0.0)",
    "pywebview (>=5.0.0,<6.0.0)",
    "pythonnet (>=3.0.0,<4.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pre-commit = "^4.2.0"
pytest = "^8.4.1"
pytest-timeout = "^2.3.1"
types-pyyaml = "^6.0.12.20250516"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (model loading, large indexing)",
]
addopts = "-m 'not slow' --timeout=60"  # Skip slow tests by default, timeout after 60s
timeout = 60  # Default timeout for all tests


